Move compiler classes to deprecated
Move CompileApp, CompileDependencies, and CompileDiScripts from src/Compiler/ to src-deprecated/Compile/ as these will be replaced by Ray.Compiler module functionality.
