+++ src-deprecated/Compile/CompileDependencies.php
@@ -2,10 +2,12 @@
 
 declare(strict_types=1);
 
-namespace BEAR\Package\Compiler;
+namespace BEAR\Package\Compile;
 
+use;
+use;
+use BEAR\Package\Compiler\NewInstance;
 use Ray\Di\AbstractModule;
-
 use function array_keys;
 use function assert;
 use function is_int;
@@ -13,6 +15,7 @@ use function sort;
 use function strpos;
 use function substr;
 
+/** @deprecated Use CompiledInjector */
 final class CompileDependencies
 {
     public function __construct(
